+++ Assets/Mirror/Transports/SSL/Certificate.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+namespace Mirror
+{
+    public class Certificate : MonoBehaviour
+    {
+        public SSLSettings sslSettings;
+        public CertificateSettings certificateSettings;
+    }
+}
