+++ Assets/Mirror/Transports/SSL/CertificateSettings.cs
@@ -1,44 +1,11 @@
 using System;
 using System.IO;
-using System.Net.Security;
-using System.Net.Sockets;
-using System.Security.Authentication;
 using System.Security.Cryptography.X509Certificates;
+using UnityEditor;
 using UnityEngine;
 
 namespace Mirror
 {
-    public class Certificate : MonoBehaviour
-    {
-        public SSLSettings sslSettings;
-        public CertificateSettings certificateSettings;
-    }
-
-    [Serializable]
-    public class SSLSettings
-    {
-        [Tooltip("Enable SSL for mirror transport components? (default: false)")]
-        public bool SSLEnabled;
-
-        [Tooltip("Protocol to use for ssl (default: TLS 1.2)")]
-        public SslProtocols SSLProtocol = SslProtocols.Tls12;
-
-        public Stream CreateStream(NetworkStream stream, X509Certificate2 certificate)
-        {
-            SslStream sslStream = new(stream, true, AcceptClient);
-            sslStream.AuthenticateAsServer(certificate, false, SSLProtocol, false);
-
-            return sslStream;
-        }
-
-        // Always accept client
-        private bool AcceptClient(object sender, X509Certificate certificate, X509Chain chain,
-            SslPolicyErrors sslPolicyErrors)
-        {
-            return true;
-        }
-    }
-
     [Serializable]
     public class CertificateSettings
     {
@@ -128,4 +95,10 @@ namespace Mirror
             return true;
         }
     }
+
+#if UNITY_EDITOR
+    // [CustomPropertyDrawer(typeof(CertificateSettings))]
+    public class CertificateSettingsDrawer: PropertyDrawer
+    {}
+#endif
 }
